requirements.txt
asgiref==3.5.2
Babel==2.10.1
dj-rest-auth==2.1.11
Django==4.0.4
django-cors-headers==3.12.0
django-money==2.1.1
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
py-moneyed==1.2
pytz==2022.1
sqlparse==0.4.2
